// src/app/core/api/generated/model/order-response.ts
/**
 * Bloom Studio API
 * REST API pour la gestion de la boutique florale Bloom Studio
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: bonjour@bloom-studio.fr
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderItemResponse } from './order-item-response';


export interface OrderResponse { 
    id?: number;
    orderNumber?: string;
    userId?: number;
    userEmail?: string;
    items?: Array<OrderItemResponse>;
    status?: OrderResponseStatusEnum;
    deliveryAddress?: string;
    deliveryZip?: string;
    deliveryCity?: string;
    deliveryMethod?: string;
    deliveryNotes?: string;
    subtotal?: number;
    shippingCost?: number;
    total?: number;
    createdAt?: string;
}
export enum OrderResponseStatusEnum {
    PAYMENT_PENDING = 'PAYMENT_PENDING',
    PAID = 'PAID',
    PENDING = 'PENDING',
    CONFIRMED = 'CONFIRMED',
    PREPARING = 'PREPARING',
    SHIPPED = 'SHIPPED',
    DELIVERED = 'DELIVERED',
    CANCELLED = 'CANCELLED'
};